Initial checkin.  I created this by combining the XBoard 4.2.6 and WinBoard 4.2.6 source distributions, eliminating duplicate files, renaming the Windows MSVC makefile, removing superfluous files, and removing files from separate projects that were included in the WinBoard 4.2.6 binary installer package (gnuchess4, gnuchess5, ICC timestamp, and FICS timeseal).  The file contents are unchanged.
This version of the project does not compile as-is.  The makefiles
need changes to work with the new directory structure.  It's here only
as a base reference point.

The initial CVS version is set to 2.0 to avoid overlapping with the
RCS version numbers that were used in past releases of the project.
They all were in the 1.x series except cmail, which was in the 3.x
series.