corrected 'animate dragging'
I finally figured out how the 'animate dragging' works in XBoard,
so that I could add code to do the unstacking of multiple pieces
in the holdings, rather than always leaving an empty square while
you drag them. (And then suddenly re-growing the remaining pieces
as you make the drop.)  It does not look as fancy as in WinBoard
yet, because the counts still only decrease when you release the
piece, rather than when you lift it. But this is not nearly as
disturbing as leaving empty holdings (if only temporarily) when
you still have pieces there.