added the Action-menu items that allow the user to adjudicate ongoing games in Two-Machines mode.
This was an already existing backend function from Winboard_x; I only had to create a menu item for it to be called. Due to an oversight I had forgotten to do that so far. (Logically, I should have done it in the 4.3.5 release, where I updated the XBoard menus to conform to the WinBoard menus.)

I tested them, and they seem to work. (As expected.)