bugfix and enhancement in autoKibitz mode
changed the format of the autoKibitz: score/depth in stead of depth/score, as in PGN. I also supress the ICS time for the move now in the PGN if extended info is available.

When enhancing the functionality of the autoKibitz command (using the
opponent engine kibitz for the saved PGN) I discovered that the xengineoutput.c
in our tree was obsolete, and contained a routine of which the arguments
were no
longer compatble with the call from backend.c, and caused segfaulting. I
copied in
the  corresponding routine in wengineo.c, which was compatible. That the
compiler
did not flag this is a sign of badly organized code, I suppose...

The enhancement itself affected backend.c.