Refined stalemate adjudication in Suicide, some cleanup
Refined stalemate adjudication in Suicide, which seems to be not always won
for the stalemated side, but for the side with the smallest number of pieces.
Cleaned up the adjudication code in the process. (backend.c moves.c, moves.h)
Also includes a tiny change in the -autoKibitz format, as requested on
Talkchess.