two shogi related bugfixes
I discovered a bug in the move generator for the Shogi Dragon King. I
guess it went unnoticed so far because this is a promoted version of the
Rook, which hardly occurs in games. (moves.c)

There was also a problem with the piecesfix I sent you before; Shogi
chosen as initial variant did use wrong piece symbols. (I had tested this
on gothic, but there it happened to work for a different reason, unique to
gothic.) This affects xboard.c.