fix for Animation of castling moves in Chess960
Animation of castling moves in Chess960 turned out to sometimes make the
king disappear from the display. For the time being I patched this in
xboard.c to suppress animation of FRC castling moves altogether. As FRC
castlings internally are represented as KxR, the to-square does not make
sense anyway, which made the animation look very weird even when it led
not to a king disappearance act. And the Rook move of a castling is never
animated.