integrated jaws into winboard
I integrated the JAWS patch now into WinBoard in such a way that the
JAWS and standard version can use the same resource file. This new
winboard.c contains the improved dialog layot of the JAWS version,
but none of the menus and JAWS-specific acclerator keys. The latter
are now generated at run time by code in jaw.s. As a result jaws.h
could be cancelled, as there the JAWS-specific defines are no
longer needed outside jaws.c (now the .rc does not use them), and
is incorporated in it.