Updated cmdline makefiles to be able to produce JAWS binaries.
This breaks support for MSVC 4.x.  It may still be possible to
use that compiler if you obtain the latest SDK from Microsoft.
MSVC 6.0 still works, but I had to change the target windows
version to 0x500 which causes warnings about building beta
applications.  This is all due to the use of MENUBARINFO data
type in the jaws binary.  You can reduce the WINVER to 0x0400
safely if you are not building a JAWS binary, and MSVC 4.x
will work fine.