fix premove for winboard
This might fix the illusive premove problem. In 4.2.7 the code was not properly
indented, and this led to an unbalanced braces error that I then fixed the
wrong way.
Click-click premoves did only work after clicking the toSquare twice. It
might be
that they actually did work after clicking it once, but that the highlights
were not
displayed (although set) for lack of a display refresh.