make xboard usable with UCI engines
I made a small patch in xoptions.c (where the corresponding Engine Settings
dialog was for XBoard), to make it at least usable for UCI engines: I now
divide the options over upto 4 columns, if there are too many to fit into
a single column. This seems to work. Still looks very ugly, but it is
functional.