fixed enable zippy in configure script and added some checks for OS X
make tried to compile zippy even when --disable-zippy was given. Also fixed a bug in xboard.c where #ifdef instead of #if was used to test for ZIPPY (it's either 0 or 1, so we need #if).

Also added some special checks for OS X, not sure if these will do it, but it's a start;)