updated installe script
I re-organized the doc files in the install, so they now all go in the
WinBoard\doc folder
(and none in the root). I so included the COPYING and COPYRIGHT files there.
This required some updates of the links in some html files, and of the menu
link
to README.html in the install script. I also fixed an error with the
current directory
in the menu short-cuts of some engines (which started in the fonts folder, and
could not find the opening book there...).