workaround for error message for missing pieces used in variants and initstring bugfix
error messages showed for non-existing pieces showed up for old user supplied pixmaps. Loading a pixmap of the king for the missing pieces now.

Another bad bug surfaced: the -initString argument does not work in any
XBoard version! (Including 4.2.7b) The escape sequences for linefeed in
the string are not understood: -initString "new\nrandom\n" actually sends
the \ and n to the engine (which then does not recognize the line as a
valid command)!

So I also added a patch to expand escape sequences in the InitString and
ComputerString command-line options.