updated polyglot protocol
I added a new option type "-reset" to the protocol for the new option feature.
4.4.0 does not implement it yet, but Michel wants it for future Polyglots,
and I think it is best to update the protocol in one big swoop. So I want to already
include it in the specs now. (engine-intf.html)

I made xboard's response to an unknown option type compliant with the new protocol specs.