fix for keepalive and chat window
I made a fix to the keepAlive option so it does not send the date command
to the ICS after every move. This required changing ScheduleDelayedEvent()
so that it does not execute the event if a new identical one gets scheduled.
I also added the option to XBoard. Affects
xboard.c (DOS!) and winboard.c

(This also cures the regression of the winboard.c copyright patch)

I extended the chatbox patch to also listen to channel tells, and divert
them to a chatbox for that channel number. Affects
backend.c and wchat.c

That should make those patches fully operational. No docs for either of them
yet, though.