bugfix for protocol extensions: egtbpath -> egtpath
A new and very bad bug was discovered in the implementation of the protocol extensions:
XBoard sends a command "egtbpath" to the engine in stead of "egtpath" from the protocol specs.
After discussion on WB forum it was decided that egtpath was better than the egtbpath I originally
had proposed, and I changed it accordingly, but I had not realized that this string occurred TWICE
in backend.c, depending on if matched the -defaultPathEGTB of -egtFormats option.
I only changed the first occurence.