new forceIllegalMoves option
The -forceIllegalMoves option uses setboard or edit to force a position when an engine refuses the move leading to it by an "Illegal Move" message.
Especially for engines using the edit command this was a bit tricky, because of the a2a3 kludge to avoid using the white and black commands.