Updated navigation accelerators, fixing ICS problems.
Recent changes to the navigation accelerators were in the wrong
accelerator table.  Since the goal was to correct non-JAWS versions
back to 4.2.7 behavior, I updated it to how 4.2.7 did it, but still
retained the #ifdef JAWS conditional compilation.

On an ICS, the navigation keys took control away from the ics
client to navigate the game.  The update requires the ALT key to
be pressed, but it works identical to 4.2.7, not causing an
unexpected focus back to the board window when you're just editing
text.