fix joining of lines split by ICS
This patch fixes the joining of lines that were split by an ICS, so that adding
a space at the break point is done only when there was not a space already.
(Some ICS leave a space at the end of the broken line, others do not.)
An extra space would interefere with board12 in Gothic Chess, which is so
long that the ICS breaks it, and the WB parser could not handle a double
space within the board.