new mousehandler to correct for premove and promotion popup
this is the new mouse handler for XBoard, which should fully implement
premove, including promotion popup, and yet suppress the promotion popup
on illegal moves that are not premoves. It is almost completely moved to
the back-end now, so that WinBoard will be able to use it as well. I just
must hook it up to winboard.c, in particular harmonize the call-backs of
the mouse driver into the front-end for grabbing and releasing pieces for
dragging, to make sure they have the same names and arguments.