fix for bug #27667: PV line missing in analysis window, part 3
"3. Wish for Engine state string (right to the icon): Can you add the actual search depth in analyzing mode? A change like "

I also added the requested display of depth of the analysis move in the
EngineOutput window. I guess a similar change could be made to wengineo.c (as this is in the
code they share in common), but no one ever requested that in the 5 years
since Alessandro wrote it. So for now I leave it alone.