fix for bug #8847: moving backward while examining on FICS not reported to engine
I moved Daniel's code that backs up the engine when a bord with a lower number is received to a place where it is always done in zippy mode, rather than only when the engine is analyzing.

Apart from taking back moves, there was also the case where the
user moves forward by more than one move (usually immediately to
end-of-game). In this case a new movelist is fetched from the ICS
to recover the moves that were skipped. But this assumes that the
engine is at the start of the game and feeds all these moves to the
engine. So the engine had to be reset in this case to stay in sync.