holdings update and regression fix
The holdings-update patch was not yet satisfactory, because FICS seems to send
holdings both before and after the board! And those received before did overwrite
the holdings received after the previous board. I know tried to fix that in a kludgey
way, by hiding a flag in the board that tells if the holdings have already been written
by the ICS or not, and ignore new holdings when they are. (Except in bughouse,
where holdings updates cannot be tied to a particular board and multiple updates
between boards are normal. (I hope?).)

I also undid a regression in my previous fix for the minor graphics problem,
and solved it in a slightly different way to prevent flicker.