Silence warnings when compiling 64-bit xboard
Some harmless warnings were caused by 64-bit mode having 32-bit
int but 64-bit long and pointer.  Add casts (etc.) to silence them.
This might require adding a definition of intptr_t (a signed int
type that is the same width as a pointer) on old systems that don't
have it.