refactoring evalgraph code
This is the refactored EvalGraph code: the back-end part is separated off,
and put in a file evalgraph.c, which in the future can also be used with XBoard.
Only the WinBoard front-end code is left in the file wevalgraph.c. A new
header file defines the variables and routines they share.