1st step for moving option parsing from winboard to the backend
This is a preparatory step in moving option parsing of WinBoard to the back-end.
The table of command-line options now contains a fifth field, holding the compiled-in default.
This obviates most of the front-end routine InitAppData, which now is replaced by a table-driven
initialization routine SetDefaultsFromList(), which could be back-end.