2nd step for moving option parsing from winboard to the backend
This is the next step in option-setting refactoring: All options that are
defined in XBoard are now included in the WinBoard ArgDescriptor list.
The handling of sound option has been harmonized: the sound names
now all initially go to the appData array, (as in XB) in stead of directly to the
WB-specific variables, and LoadAllSounds now takes care of the copying.
A new routine ExportSounds copies them back before saving settings
(which, in XBoard, could be a no-op).

The definition of defaults for the XB-only options has been moved to common.h.