same argDescriptor parsing for Xboard and Winbaord
This patch makes XBoard and WinBoard use the same argDescriptor
list of options, and (largely) use the same code for parsing and saving them.
This is acheived by moving this code to a new file args.h, and #including this
in both winboard.c and xboard.c (so that tiny differences, e.g. in default
values in the table, are decided by the context). XBoard has new menu items
"Save Settings Now" and "Save Settings on Exit" in the "Option" menu.
XBoard uses /etc/xboard/xboard.conf as default settings file.
A new option, -saveSettingsFile, can specify where settings should be saved,
without parsing the file. (So that it also works if the file does not yet exist.)