Indicate squares a lifted piece can legally move to
Under control of the new option -showTargetSquares when
-highlightDragging and -testLegality are on, the squares a piece can
move to are marked by fat dots in the highlightColor (non-captures) and
premoveHighlightColor (captures), as soon as you grab it for dragging.
In variants with mandatory capture, the capture target of other pieces
is marked with the highlightColor (as non-captures are then not allowed).