Extensive bugfix of -autoKibitz
Suppression of printing of diverted kibitz messages failed when they
extended over more than one read. This could even lead to such
messages being erroneously saved as comments, which then lead to
crashes. The suppression mechanism now waits for a prompt, in stead of
the end of buffer. This required the printing of any leftover in the
buffer to be deferred until the next read completes it.
The number of people kibitzed to is now printed.