Add variant Makruk
Thai Chess, played by 2 million world-wide. Required pawn promotion to
happen in a zone, and the Alfil to move as Shogi Silver, a new opening
array, and unusual Pawn placement. Because of the latter, I consider it
a setup position, so a FEN will always be sent to the engine.