Port game-list filtering to XBoard
Added a text edit and apply button to the XBoard game-list dialog, and
only have it print those lines that match the filter. Translate index in
list to index of game in file when a game is selected, and index in file
to index in list when a line is highlighted. Duplicates some code
between xgamelist.c and wgamelist.c, which should be moved to back-end.