Refactoring of adjudication code
Put all the rule-based (as opposed to score-based) adjudications in a
separate routine, so they can also be called on user moves. This
requires sending the drawing move to the opponent engine to be made
conditional on existence of the latter. Execution of the code is no
longer dependent on the condition TwoMachinesPlay, but actual game
terminations are now all restricted by an alternative condition (no ICS
mode).