Send FICS atomic claim to ICS if move creates draw after offer
The FICS format "draw MOVE" is used to send the move to the ICS  when
the user enters a move that adjudicates as a claimable draw, and he
offered a draw before the move. For the benefit of ICS that do not
understand this format, the plain move is sent afterwards.