Extensive bugfix of promotion pseudo-popup
With holdings the pieces were offsetted vertically in stead of
horizontally, and in shogi the pieces were not put on the promo board at
all. Also the choice in shogi had to be translated into + or =, while in
crazyhouse the promoted forms did not translate to chracters well, so
now the basic types are shown in the pseudo-popup instead.