Repair in WinBoard what the XBoard PromoDialog broke
Because the back-end dummy PromoDialog was removed, and its function
taken over  by a front-end working version in XBoard, WinBoard needed to
supply its own front-end. So the old back-end dummy was moved to
winboard.c. The code to suppress some upclicks, which is now only needed
in WinBoard because it still uses pseudo-popups, was put in the mouse
handler.