Fix bug in edit-position of holdings
The holdings were effectivelly taken one too large, so that a promoted
shogi Pawn moved to, or created in the holdings was not demoted to Pawn,
but put in the dark (which you could see by the holding count if you did
it twice). The overflow tests in these two cases now have been corrected.