Fix u64 format for cygwin
With a -mno-cygwin compile %llu and %lld are definitely not understood
by scanf and printf, although they are by the compiler. In stead, %I64u
seems to work. Put it in under control of a switch USE_I64, which is
defined only in the gcc makefile when MINGW is used.