Implement SeekGraph in XBoard
For now only controlled by left mouse clicks on board. New options
-seekGraph true|false and -sg to switch this feature on.

Facilitate selection of coinciding seek ads

Seek-ads that are viewed but not answered get a distance penalty on the
up-click that makes their selection more difficult next time, so that
the ad they covered gets a (better) change. The penalty ages away in the
course of 5 clicks.

Slightly offset Seek-Graph dots of different color

Dots of different color are now displaced horizontally be 3 pixels, to
prevent them from completely covering each other. (Required the axis
labels to be moved a little too, to not be covered by the 1-min ads.)

Auto-show seek-ad texts in WinBoard

Hovering the mouse above a Seek-Graph dot will make the text belonging
to it appear in the message field without the need to press a button.
Required SeekGraphClick() to be called from the WB mouse event handler,
with an extra argument 'moving' to make the distinction with a real
click.

Show seek-ad text on mouse hovering (XBoard)

The Eventproc is also set to trigger on PointerMotion, and this event
then calls SeekGraphClick with motion=true to handle it.

Put poetic name of wilds in seek-ad text

Use XBoard variant names, except for 'loadable' and 'fischerandom'
(becomes 'setup' and 'chess960'), but leave number (to distinguish wilds
that transate to same XBoard variant).