Use right mouse button to view seek ads
Right down-click does the same as left down-click, right up-click
simulates a left up-click off board (so it misses all dots).

Use right-click miss to refresh Seek Graph

The 'moving' parameter had to be made into an int for this, to indicate
right clicks with code 2. Right press now already pushes the clicked ad
to the back, so that the up-click will display an underlying one.

Bugfix pushing back seek ads on right-click

On a right button press the previous in stead of currently hit ad was
receiving the penalty.

Indicate seek-ad multiplicity

If more than one seek ad is inside the capture range of the mouse
pointer, an exclamation point is printed before the seek-ad text.

Some fixes to make seek-ad suppression work in WinBoard

In WinBoard the z-depth was immediately erased by redrawing the graph;
the clearing of zlist had to be moved from PlotSeekAd() to AddAd()
(where it really belongs). Also aging of z-depth is now suppressed on
mere pointer motion.