Merge gamelistopt sources into gamelist source files
The files xgamelistopt.c and gamelistopt.c are merged with xgamelist.c
and gamelist.c, respectively. The WB front-end for the dialog already
resided in winboard.c, so there was little need to move it to wgameist.c
file. Makefiles adapted accordingly. Unsuitable key bindings were
removed. The low-level front-end prototypes are now defined in in
frontend.h.