Send continuation lines to chat box they belong to
The variable savingComment is used as a kludge to remember where the
comment should go (i.e. which chatbox, or elsewhere), so that when a
continuation line is matched, chatPartner (which was reset at EOL) can
be restored to its original value. Also put space after handle of talker
in channel Chat Windows