Bugfix stale first-click
When the opponent resigned after the user had cicked a from-square,
clicking the from-square for the first move in the next game would
produce an "illegal move" message, because it was considered a to-click!
We now set fromX and fromY to invalid in GameEnds().
Should we also reset them on mode changes?