Observe a game in the background while playing
A new (persistent) option -backgroundObserve true|false is added. When
set, observed boards received while playing a game will not trigger a
game switch, but will be ignored except for storing in a dedicated board
buffer. Pressing the right mouse button will cause the board to be
displayed, until we release it. Time, strength and stm are always
displayed in the message window.