Allow WB Chat Box to be dedicated to shouts and 'it'
By putting 'shouts' in the handle field, a chat window will capture all
shouts and c-shouts, and send messages typed into it as shouts.
The bracketing of the shouter name now indicates the type of shout.
The handle for WHISPER Chat Boxes has been changed to 'whispers', to
make it easier to pop them up by clicking the console.