Fix copying of WB Chat Window contents
All line breks were lost on copy-pasting the contents of a chat window.
Adding CR to each outputted line in the front-end (wchat.c) solves this.
(Tell messages sent to a handle (which are not echoed by the ICS, and
thus have to copied to the chat box directly) also need CR!)