Allow ICS context menu to pop up a Chat Box on clicking handle
A command "chat" defined in the WinBoard icsMenu string will cause
ChatPopUp() to be called when it is clicked, with the clicked handle
(name) in the ChatPartner field, and focus on the input field.
ChatPopUp() had to get a parameter (the handle) to make this possible.
If that parameter is NULL (as when it is called from the Mode menu),
focus will still start on the (empty) ChatPartner field of the box.