Pop up ICS text menu with default item under mouse pointer
The menu popup is migrated to the down-click (was up-click), so the
up-click can select the item. Some trickery is needed to fool the menu
into thinking that the mouse moved, so that a moveless up-down click in
the ICS text field selects the default item (the upper-right one).
Also allow dummy item in ICS text menu: The command "none" in the ICS
text menu will now be treated as a grayed-out (disabled) entry, so it
can be defined in the position underthe mouse pointer to prevent the
up-click from doing anything, while keeping the menu open. Recommended
for use with an entry of spaces.