Fix XB crash on giving keyboard focus to non-text widget
In the Engine-Settings dialog SetFocus() could be called for non-text
widgets, and this made XawTextSetInsertionPoint() croak. Setting the
insertion point to the end in SetFocus() was a bad ide anyway, as it
interfered with copying from the engine-output window, which would
uncontrollably scroll on getting focus. It is solved now by setting the
insertion point already on creation of the text widgets.