Add Annotate item in Step menu
This does the same as Revert, except that it adds the destroyed
variation as comment to the original movem in PGN variation format.
RevertEvent now has an argument to indicate if it should add the
comment, or not. Comments in the abandoned variation are included in the
varriation command to the main line. (The exact layout of the variation
comment that results leaves still to be desired.)