Support playing through PGN variation comments
Right-clicking a variation line in the comment dialog will shelve the
current variation, and parse the comment to replace it. This uses
ParsePV() in an alternative way, for which the latter had to be
enhanced: it was made resistent to move numbers and comments inside
the PV. It now has an argument to indicate if such comments should be
stored with the moves. The Comment Popup is closed on revert, to
make sure no comments to moves that are destroyed, containing
variations on a non-valid line, can keep hanging around.