Let yy_text determine progress of PV parsing
The parsed unit yy_text was not available after ParseOneMove(), because
the buffer switch in yylexstr() apparently destroys it. To solve that,
yylexstr() now is equiped with an extra buffer argument, where it copies
yy_text before the switch, so the caller, ParsePV(), can see what was
parsed in yy_textstr. This is used to update the PV pointer, but also to
save any parsed comments (when ParsePV is used for parsing a PGN variation).