Seek-Graph bugfix: disappearing output in ICS console
Suppression of output to the cosole was handled fundamentally wrong: it
was not enough to set next_out to after the unit that should be
suppressed, as there can be a backlog in printing of previous units that
came in the same packet. So this has to be flushed first, or it will be
lost. -autoKibitz probably got away with that, because it comes in a
separate packet. But seek ads are often suffixed to other output, and
then that output disappeared! Now we flush everywhere before suppressing.
We now totally rely on the prompt to restart the printing, so that
acknowledgements following the echo of the sent message are also
suppressed. Also suppress LF before diverted chat messages.
  Also fixed is the 'mamer bug' in Chat Boxes, due to false parsing of
a "* (*): " pattern in tell messages, because of a missing continue.