Use side-by-side boards to display background game
A game observed in the background (i.e. if -backgroundObserve true) will
be displayed right of the normal board in the same window if the new
(persistent) Boolean option -dualBoard is also set.
Very flakey hack, especially in XBoard, where clocks and messsage
field all get the wrong width (and resist adjustment in the strangest
ways). The clocks of the partner game are not yet displayed, and
disaster happens when the background and foreground games do not have
the same board format.