Add option Display Logo in WB general-options menu
Changing the option requires InitDrawingSizes to be called for
re-formatting the board window. Switching it off requires clearing of
the cps->programLogo fields to take effect immediately. Switching it on
only takes effect the next session, as we will need some refactoring to
redo the figuring out what logos to display.