Redraw second board on expose events
The -dualBoard option now does refresh the second board after another
window uncovers it. To acheive this, some data structures remembering
what is displayed had to be duplicated. (lastBoard and damage for XB,
lastReq, lastDrawn, lastReqValid and lastDrawnValid for WB.)
WinBoard now uses clipping also for second board. This required the
clips for the destination to be translated horizontally!