Set keyboard focus at startup to board window
This makes XBoard sensitive to the keybindings also when the cursor is
outside the board window. This focus seems to persist indefinitely: each
time the board window gets selected, the key bindings become activem
wherever the mouse pointer is. Downside: the mouse pointer is no longer
an arrow outside the board window, but becomes a text-insert tool there.