Allow loading of PGN variations in XBoard
Right-clicking variation comments in the comment-window was implemented
through a translation. The right button is now bound to routines that
set both the end and start of the selection, as setting a new start does
not seem to work if the previous select-start was not completed by a
select-end or extend-end. Finally the main handler uses the selection
cursors to know what was clicked.