Suppress background observe for boards with own game number
This is really a work-around for an ICS bug. The Variant-ICS sent the
initial board of a game in a move list with relation = 0, which means
'observed'. So XBoard was thinking it was observing its own game, and
switched to dual board. It seems harmless to suppress such a switch for
all boards of your own game, however, which is now done.