Fix info lines being used as normal thinking output
The new standard of using thinking output with time = nodes = score = 0
as info messages for printing in the Engine-Output window was not
understood everywhere, and led to such info lines being used for
autoKibitzing and PGN annotation. The programStats are now only updated
from thinking output received from an engine when it is not an info line.