Alter XBoard key bindings to mimic WinBoard
A lot of new bindings were added, and four menuProcs had to be registered
to be able to add bindings to them. The use of the control key as
temporary take-back for the last move conflicted with the use of Ctrl
keystrokes, and this function was migrated to '.'. The iconify function
had no shortcut in WinBoard, but to get rid of letter bindings it was
migrated to '-'.