Cut board squares out of texture bitmap more cleverly
When the bitmap is larger than the entire board, the squares are now cut
out of the center of the virtual squares from subdividing the bitmap, so
that they stay centered on the grid of a Xiangqi board bitmap. If the
board size exceeds that of the bitmap, the old algorithm is still used,
as the centering would make the edge squares exceed the bitmap, and
cause a black rim around the board.