Started on the preference dialog
This should be quite nice... everything will be defined in glade. We need a struct with all preferences in xoption.c which we will use to loop over. A callback on each setting will go into a switch where a copy of appData is changed. That way we can OK the changes at the end.. Alernatively we could have the changes live and reset if on cancel. Anyway, this commit only has the basics. More to come in the next days