Remove font settings from master settings file
xboard.conf should not contain general font settings (not specific to
one board size) with specified point size. In fact it should not even
specify sizable fonts, as the new system would make specification of a
general font overrule the size-specific settings in the .xboardrc file,
so the persistence of fonts would be completely broken.