Fix parsing of SAN shogi promotions
A missing promotion character defaults to promotion in the move type,
but the promotion character was not set when this happened in variant
shogi. And it is especially important there, as the promotion character
'+' will always be missing, as the parser will take it to be a 'check'
symbol, and strip it from the move.